+++ src/CS/CS_algorithm.cpp
@@ -75,7 +75,7 @@ void CS_algorithm::minimize() {
     } else {
       n_sc = 0;
     }
-    if (n_sc > m_algo_config.getNMaxIterationsSameCost()) break;
+    if (n_sc > (int)m_algo_config.getNMaxIterationsSameCost()) break;
 
     // Update progress bar
     progress_bar.increment();
