+++ src/HS/HS_algorithm.cpp
@@ -76,7 +76,7 @@ void HS_algorithm::minimize() {
     } else {
       n_sc = 0;
     }
-    if (n_sc > m_algo_config.getNMaxIterationsSameCost()) break;
+    if (n_sc > (int)m_algo_config.getNMaxIterationsSameCost()) break;
 
     // Update progress bar
     progress_bar.increment();
